+++ crates/nixci/src/command/mod.rs
@@ -1,5 +1,0 @@
-//! CLI commands for nixci
-pub mod core;
-pub mod gh_matrix;
-pub mod run;
-pub mod run_remote;