+++ crates/omnix-cli/src/command/mod.rs
@@ -1,6 +1,7 @@
 pub mod ci;
 pub mod completion;
 pub mod core;
+pub mod hack;
 pub mod health;
 pub mod init;
 pub mod show;