+++ crates/omnix-health/src/check/mod.rs
@@ -1,0 +1,9 @@
+//! Individual Nix checks
+pub mod caches;
+pub mod direnv;
+pub mod flake_enabled;
+pub mod max_jobs;
+pub mod min_nix_version;
+pub mod rosetta;
+pub mod shell;
+pub mod trusted_users;