+++ doc/src/history.md
@@ -4,10 +4,13 @@
 
 ### Enhancements
 
+- `om hack`: New command
 - `om init`
   - Initial working version of `om init` command
 - `om health`
   - Display Nix installer used (supports DetSys installer)
+  - Display information in Markdown
+  - Remove RAM/disk space checks, moving them to "information" section
 - `om ci`
   - Support for remote builds over SSH (via `--on` option)
   - Support for CI steps